requirements.txt
requests 
beautifulsoup4 
sentence-transformers 
transformers 
streamlit
flask
faiss-cpu
sentencepiece

